fix integration with SI tab, move SI fetch method into tab state store
